removed some methods from Material

+++ common/src/main/java/muramasa/antimatter/material/MaterialTypeFluid.java
@@ -28,6 +28,6 @@ public class MaterialTypeFluid<T> extends MaterialType<T> {
     }
 
     public interface IFluidGetter {
-        FluidStack get(Material m, int amount);
+        FluidStack get(Material m, long amount);
     }
 }
